feat: 패티를 구워버렸다
- AttachPoint (gameObject) : Tag = Stove 지정, Stove (Script) 추가, Box Collider 의 Is Trigger = True
- Rare Patty (Prefab Asset) 생성 : Patty Controller (Script) 추가, Tag = patty 지정]
- Patty Controller (Script) : isStart, isBaking 변수 추가 및 패티 굽는 로직을 update 에 설정
- Stove (Script) : Collider 이용하여 물체 충돌 감지 (OnTriggerEnter, OnTriggerExit)
