+++ Assets/Scenes/yourzinc/Patty/Patty.cs.meta
