+++ Assets/Scenes/yourzinc/Patty/PattyController.cs.meta
