+++ Assets/Scenes/yourzinc/Patty/Rare Patty.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 854b54e60a0a9b447a92682d85699eea
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
