+++ Assets/Scenes/yourzinc/Patty/Stove.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 113f3e6bcc3ee994898ba93ba9470668
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
