+++ Assets/Scenes/yourzinc/Patty/Stove/Stove.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cf08595bca2fa7b4896e67e9f0af99d3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
