+++ ProjectSettings/TagManager.asset
@@ -3,7 +3,9 @@
 --- !u!78 &1
 TagManager:
   serializedVersion: 2
-  tags: []
+  tags:
+  - patty
+  - Stove
   layers:
   - Default
   - TransparentFX
